Refactor tool into separate modules
Extract the monolithic `main.rs` into dedicated modules for
CLI (`cli.rs`), commands (`commands.rs`), the flash device
abstraction (`device.rs`), and transport backends (`transport.rs`).
